backend/requirements.txt
fastapi
uvicorn

pandas
numpy
scikit-learn
joblib

pdfplumber
python-docx

opencv-python

deepface
mtcnn

librosa
soundfile

openai-whisper

transformers
torch
sentencepiece

python-multipart